Refactoring move topology_builder into lib
