+++ lib/api/rest_api_base.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 require 'grape'
-require 'logger'
+require 'lib/netomox_exp'
 require_relative 'helpers'
 
-# NetomoxExp Application (REST API server)
 module NetomoxExp
   # Rest api base class
   class RestApiBase < Grape::API
     format :json
-
-    helpers NetomoxExpHelpers
+    logger NetomoxExp.logger
+    helpers Helpers
 
     helpers do
       # @return [Logger] Logger
@@ -18,28 +17,5 @@ module NetomoxExp
         RestApiBase.logger
       end
     end
-  end
-
-  # module common logger
-  @logger = RestApiBase.logger
-  @logger.progname = 'netomox-exp'
-  @logger.level = case ENV.fetch('NETOMOX_EXP_LOG_LEVEL', 'info')
-                  when /fatal/i
-                    Logger::FATAL
-                  when /error/i
-                    Logger::ERROR
-                  when /warn/i
-                    Logger::WARN
-                  when /debug/i
-                    Logger::DEBUG
-                  else
-                    Logger::INFO # default
-                  end
-
-  module_function
-
-  # @return [Logger] Logger
-  def logger
-    @logger
   end
 end
